Render pages based on state
