+++ context/context.js
@@ -1,0 +1,41 @@
+import React, { useReducer } from 'react'
+
+const reducer = (state, action) => {
+  let result = {
+    ...state
+  }
+
+  switch (action.type) {
+    case 'change-page':
+      result.route = action.payload
+      return result
+    default: 
+      return result
+  }
+}
+
+const intitialResponseState = {
+  route: 'home'
+}
+
+const changePage = (string) => {
+  return {
+    type: 'change-page',
+    payload: string
+  }
+}
+
+const initState = () => intitialResponseState
+const Context = React.createContext({})
+
+function ContextProvider(props) {
+  const [state, dispatch] = useReducer(reducer, undefined, initState)
+
+  return <Context.Provider value={{ state, dispatch }}>{props.children}</Context.Provider>
+}
+
+export {
+  Context,
+  ContextProvider,
+  changePage
+}
